Backend/requirements.txt
# Afya Mawinguni Backend Requirements
Flask==2.2.3
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
Flask-JWT-Extended==4.4.4
Flask-Bcrypt==1.0.1
PyMySQL==1.0.2
mysqlclient==2.1.1
python-dotenv==1.0.0
Werkzeug==2.2.3
marshmallow==3.19.0
validate-email==1.3
python-dateutil==2.8.2
gunicorn==20.1.0
Pillow==9.4.0
